+++ cut-loader.cordova.js
@@ -113,7 +113,7 @@ function AppLoader(app, configs) {
 
     DEBUG && console.log('Resize: ' + width + ' x ' + height + ' / ' + ratio);
 
-    root.resize(width, height, ratio);
+    root.viewport(width, height, ratio);
   }
 
   return root;