+++ temp/cut.web.js
@@ -1,4 +1,0 @@
-module.exports = require('../cut-core');
-require('../cut-tween');
-module.exports.Mouse = require('../cut-mouse');
-require('../cut-loader.web');