+++ CHANGES.md
@@ -10,7 +10,8 @@
 * Source files moved to `lib` and platform files to `platform`
 * Assumed CommonJS by default
 * `lib/main` is set as npm `main`
-* extension files moved to ext 
+* extension files moved to ext
+* In cutout selector, texture name is made optional: `"[texture:]cutout"`
 
 #### v0.3.1
 * Remove `Mouse(..., captureAnyMove)` and use flag instead