+++ CHANGES.md
@@ -12,7 +12,7 @@
 * `lib/main` is set as npm `main`
 * extension files moved to ext
 * In cutout selector, texture name is made optional: `"[texture:]cutout"`
-* Cordova loader with FastContext support instead of FastCanvas
+* FastContext support added to Cordova, FastCanvas loader removed
 
 #### v0.3.1
 * Remove `Mouse(..., captureAnyMove)` and use flag instead